engine_spx2html: move emission into its own module
